Fix builds not working with multiple providers.

+++ deploy.sh
@@ -1,27 +1,24 @@
 #!/bin/bash
 
-deployment_type=$1
-branch_name=$2
-tag_name=$3
+branch_name=$1
+tag_name=$2
 now=$(date +%Y-%m-%d)
 version=$(gradle -q printVersion)
+release_branch_regex="^release.*$"
+feature_branch_regex="^feature.*$"
 
-case "$deployment_type" in
-"feature")
+if [[ "$branch_name" =~ $feature_branch_regex ]]; then
     echo "Renaming build files for feature branch..."
     rm build/libs/LGP-"$version".jar
     mv build/libs/LGP-core-"$version".jar build/libs/LGP-core-"$branch_name"-"$now".jar
     mv build/libs/LGP-examples-"$version".jar build/libs/LGP-examples-"$branch_name"-"$now".jar
     ls build/libs
-    ;;
-"release")
+elif [[ "$branch_name" =~ $release_branch_regex ]]; then
     echo "Renaming build files for release branch..."
     rm build/libs/LGP-"$version".jar
     mv build/libs/LGP-core-"$version".jar build/libs/LGP-core-"$tag_name"-"$now".jar
     mv build/libs/LGP-examples-"$version".jar build/libs/LGP-examples-"$tag_name"-"$now".jar
     ls build/libs
-    ;;
-*)
+else
     echo "Invalid deployment type"
-    ;;
-esac
+fi
